+++ expo-zustand-styled-components/src/hooks/stores/index.ts
@@ -1,0 +1,1 @@
+export { default as useAuthStore } from './useAuthStore';
