+++ expo-zustand-styled-components/src/hooks/stores/useAuthStore.ts
@@ -3,7 +3,7 @@ import { create } from 'zustand';
 import { persist, createJSONStorage } from "zustand/middleware"
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
-import { UserProfile, ViewerInfo } from '../types/profile';
+import { UserProfile, ViewerInfo } from '../../types/profile';
 import { Platform } from 'react-native';
 
 interface IAuthStore {
