--- expo-zustand-styled-components/src/stores/index.ts
@@ -1,1 +1,0 @@
-export { default as useAuthStore } from './auth';
